backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Web scraping and parsing
beautifulsoup4==4.12.2
lxml==4.9.3
httpx>=0.23.0,<0.24.0
requests==2.31.0
trafilatura==1.7.0

# Browser automation
selenium==4.15.2
playwright==1.40.0
webdriver-manager==4.0.1

# Task queue and async processing
celery==5.3.4
redis==5.0.1

# Database and storage
supabase==1.0.3
python-dotenv==1.0.0

# File handling and utilities
aiofiles==23.2.1
Pillow==10.1.0

# JWT and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0

# LLM and AI (optional - for content analysis)
openai==1.3.0

# Additional utilities that might be needed
python-dateutil==2.8.2
validators==0.22.0
asyncpg==0.29.0
